22-OOP/pyproject.toml
[project]
name = "22-oop"
version = "0.1.0"
description = ""
authors = [
    {name = "AlexanderMatveev2908",email = "matveevalexander470@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "attrs (>=25.3.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.17.0"
pyright = "^1.1.403"
watchfiles = "^1.1.0"
black = "^25.1.0"
ruff = "^0.12.5"
tomlkit = "^0.13.3"


[tool.black]
line-length = 79
target-version = ["py312"]
